Speed improvements on calendar; integer decimal to dozenal conversion now done in-script, not by system call.
Premature optimazation may be the root of all evil, but the
slowness of using system calls for conversion in the script,
since there are so many of them, drove me nuts.  So I wrote
a routine to do the conversions from decimal to dozenal
(integers only) in the script, and it feels much faster
(whether it actually is or not is another question).  Will
write a dozenal to decimal routine soon.