dozcal:  expanded astronomical and geographical options.
Can specify geographical location and timezone on the
command line; dozcal can use this information to figure out
sunrise and sunset times.  Gets latitude and longitude from
Open Street Maps a non-coordinate string is provided to "g";
if it fails, the program dies noisily.  If timezone isn't
given, the local timezone is presumed, even if this
comically mismatches the latitude and longitude.  Also
found and fixed a bug in the sorting algorithm for times; it
was cutting off any time over 84 because unshrink_time only
grabbed two digits.