dozstring:  significant updates, almost release-ready.
Can now specify custom X and E chars; like the padchar,
these must be single characters.  If you want prettier than
that, pipe the result into dozpret.  (This won't work for
padchar; but why would anyone want it to?)  Also made it
work for uncials/decimals, even when the dozenal number has
more chars than the decimal; in this case, it simply
truncates.