I've restructured the directory layout somewhat to dedupe certain files. All headers are now in dozenal/include, and the conv, dec, and doc files are now used globally rather than being duplicated in each subdirectory.
Manpages are now in dozenal/doc. Most systems I've encountered in
modern times don't include the "rman" program but do include groff,
so I'm using that instead to convert manpages to html (optional).

The build system is revamped using a master Makefile rather than the
shell scripts for install/deinstall. This provides for more flexibility
in packaging for different systems, as installation paths can be set
on the commandline for "make" as in:

% make prefix=$PWD/dozenal/usr mandir=$PWD/dozenal/usr/man

This should make the entire suite play nice with most packaging and
build systems.