Got the print_events into an arrays of strings.
Do the same for the todos, then write two new functions:
one for printing the array of events, one for printing the
array of todos.  Then the print_events and print_todos can
be called by other functions to simply fill up the arrays.

Probably need to clear the arrays prior to filling them,
too.