Repaired false error reporting in dozcal.
Also, the visual interface started doing some strange
things, with the calendar window moving the wrong amount all
the time, and the event window showing events from a day
before that reflected on the calendar.  I made no changes to
make this happen, and the fix seems wrong to me.  Still, it
works.  Keep an eye on this.