Added an option to print newline to doz().
So I had to do a lot of function changing wherever doz() is
called; this was poor planning on my part.  Now, the final
parameter passed to doz() is either 0 or 1; 1 to print the
newline, 0 to not.  This allows some dc-compatibility in
dozdc; so I added "n" to pop the top of the stack and print
no newline, along with "p" (prints the top without popping
it) and "=" (prints the top *with* a newline, popping it).