The program is now (probably) POSIXly correct.
Told gcc to be pedantic in the builds, and this revealed a
lot of unnecessary variables and the like, and even one
actual mistake.  (Fortunately, it was a mistake that didn't
effect output.)

This caused a regression in dozcal; sun times can't adjust
for time zone at the moment, because tm_goffset is not a
part of the tm struct in POSIX.  Drat.