Convert project to use `configure` script and single makefile
This commit re-organizes the source files into a single `src`
directory, and converts the project's recursive makefile structure to
use a single, top-level makefile with rules for each target binary.

The `configure` script and `Makefile` are generated using Autoconf and
Automake, from `configure.ac` and `Makefile.am`, respectively.  The
`configure` script takes the optional argument `--disable-xforms`,
which will remove `xdozdc` from the list of default targets, and will
skip the check for the XForms and X11 libraries.